Applied patches from Michael Casadevall to fix ARM issues with qreal. BUG:179060
svn path=/trunk/KDE/kdegraphics/gwenview/; revision=906327

+++ lib/documentview/documentview.cpp
@@ -238,7 +238,7 @@ struct DocumentViewPrivate {
 	qreal computeMinimumZoom() const {
 		// There is no point zooming out less than zoomToFit, but make sure it does
 		// not get too small either
-		return qMax(0.001, qMin(mAdapter->computeZoomToFit(), 1.));
+		return qMax(0.001, qMin(double(mAdapter->computeZoomToFit()), 1.));
 	}
