+++ lib/redeyereduction/redeyereductionimageoperation.cpp
@@ -106,7 +106,7 @@ inline qreal computeRedEyeAlpha(const QColor& src) {
 		axs = ramp(sat);
 	}
 
-	return qBound(0., src.alphaF() * axs, 1.);
+	return qBound(0., double(src.alphaF()) * axs, 1.);
 }
 
 
@@ -115,7 +115,7 @@ void RedEyeReductionImageOperation::apply(QImage* img, const QRectF& rectF) {
 	const qreal radius = rectF.width() / 2;
 	const qreal centerX = rectF.x() + radius;
 	const qreal centerY = rectF.y() + radius;
-	const Ramp radiusRamp(qMin(radius * 0.7, radius - 1), radius, 1., 0.);
+	const Ramp radiusRamp(qMin(double(radius * 0.7), double(radius - 1)), radius, 1., 0.);
 
 	uchar* line = img->scanLine(rect.top()) + rect.left() * 4;
 	for (int y = rect.top(); y < rect.bottom(); ++y, line += img->bytesPerLine()) {
